requirements.txt
openai
python-docx
streamlit
Jinja2
streamlit-ext==0.1.9
stqdm==0.0.5
setuptools==68.2.2
langchain
deeplake
pypdf2
python-dotenv
cohere
wave
nltk
audio_recorder_streamlit
streamlit-option-menu
streamlit-lottie
faiss-cpu
Ipython
boto3